property fixes in league models + updated mock classes

See merge request courses/2020-fall/csci-5308/group6/csci5308!32

+++ src/main/java/com/IceHockeyLeague/LeagueManager/Player/Player.java
@@ -67,7 +67,7 @@ public class Player implements IPlayer {
     }
 
     @Override
-    public int calculateStrength(IPlayerStats stats) {
+    public double calculateStrength(IPlayerStats stats) {
         return stats.calculateStrength();
     }
